IDP-1838 fix refactor connectors
