--- src/main/java/it/gov/pagopa/payment/connector/rest/paymentinstrument/PaymentInstrumentRestConnector.java
@@ -1,8 +1,0 @@
-package it.gov.pagopa.payment.connector.rest.paymentinstrument;
-
-import it.gov.pagopa.payment.dto.PinBlockDTO;
-import it.gov.pagopa.payment.dto.VerifyPinBlockDTO;
-
-public interface PaymentInstrumentRestConnector {
-    VerifyPinBlockDTO checkPinBlock(PinBlockDTO pinBlockDTO, String userId);
-}
